Move tracking epoch end outputs logic to the `EvaluationEpochLoop`
